requirements.txt
aiohttp
amcrest
asyncio-mqtt
backoff
coloredlogs
flvlib3@https://github.com/zkonge/flvlib3/archive/master.zip
hikvisionapi>=0.3.2
httpx
onvif-zeep
onvif-zeep-async@git+https://github.com/openvideolibs/python-onvif-zeep-async.git@async
semver
packaging
uiprotect==6.0.1
reolinkapi
websockets>=9.0.1
xmltodict
# skbuild  # Not available for ARM64
# opencv-python  # Not needed for Tapo module
